* Restructured common code
