--- fabric/forgero-fabric-core/src/main/java/com/sigmundgranaas/forgero/fabric/client/model/baked/strategy/CachedDefaultModel.java
@@ -1,5 +1,0 @@
-package com.sigmundgranaas.forgero.fabric.client.model.baked.strategy;
-
-public class CachedDefaultModel {
-
-}
