+++ fabric/minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/client/model/baked/BakedModelResult.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgero.minecraft.common.client.model;
+package com.sigmundgranaas.forgero.minecraft.common.client.model.baked;
 
 import com.sigmundgranaas.forgero.core.model.ModelResult;
 import lombok.AllArgsConstructor;
