+++ fabric/minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/client/model/baked/strategy/CachedDefaultModel.java
@@ -1,0 +1,5 @@
+package com.sigmundgranaas.forgero.minecraft.common.client.model.baked.strategy;
+
+public class CachedDefaultModel {
+
+}
