+++ fabric/minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/client/model/baked/strategy/SingleCachedStateStrategy.java
@@ -1,10 +1,10 @@
-package com.sigmundgranaas.forgero.fabric.client.model.baked.strategy;
+package com.sigmundgranaas.forgero.minecraft.common.client.model.baked.strategy;
 
 import java.util.Optional;
 
 import com.sigmundgranaas.forgero.core.state.State;
 import com.sigmundgranaas.forgero.core.util.match.MatchContext;
-import com.sigmundgranaas.forgero.minecraft.common.client.model.BakedModelResult;
+import com.sigmundgranaas.forgero.minecraft.common.client.model.baked.BakedModelResult;
 import org.jetbrains.annotations.Nullable;
 
 public class SingleCachedStateStrategy implements ModelStrategy {
